Test Enemy
A test enemy and NavMeshVolume have been added to the MBT_Testing scene. A basic BT has been added to this enemy, instructing it to move towards the player if they fall within a 40-unit radius of the enemy. Once it reaches a position 15 units from the player within that radius, it will stay in place. Functionality for it to shoot at the player will be added in coming pushes.