All In One Entity Architecture Implemented (Scrapped)
This commit changes the entire Entity architecture so that every Component for an Entity does not have its own state but rather relies on a singular consistent state provided by the Entity class.

While this may seem like a good idea to reduce connections, in practice it causes all logic relating to the states to become abstract and confusing, So I've opted to scrap this entire concept pending more discussion.